fix rest of macroblock metadata decoding
Turns out libwebp commit f67b5939 reordered all the things so that
they can give the iintra_mbmodes and intra_bmodes the same values
without needing the remapping table that the literal reading of the
spec requires.

Let's just use the remapping table for our code and keep the tables
in the spec.

This toggles the `#if 0` for that to `#if 1` and reverts the
previous two commits.