take intra_mbmode from libwebp instead of from the spec as well
See common_dec.h.

This isn't fallout from the previously-mentioned commit.

Maybe libwebp tries to reorder enum values only internally to
save that one lookup table. But if I keep the lookup table and
don't reorder the enum values, so far I haven't been able to
get the same output as with libwebp. Let's keep this as-is to
have a known-working state, and try to get it closer to spec
later (...or, if the needed behavior is indeed different from
what's in the spec, file a spec bug).