LibPDF: Load embedded truetype fonts for Type0 CIDFontType2
This is awkward because Type0 CIDFontType2 embedded truetype fonts
don't have a 'cmap' table and instead use mapping data in the Type0
structures.

The opentype spec says 'cmap' is required and hence our code sensibly
assumes it's there.

Since it isn't, we currently can't call a bunch of methods that read
the cmap table.

Either need to be careful to not call these methods from the PDF code,
or make the opentype code more general to pass in cmap mapping data
just so the not-spec-compliant pdf type0 "truetype" font files work :/

(For now, hacked things up a bit, else all the font loads just fail
with "no cmap")