LibPDF: Simplify CallSubr for type 1 font programs
https://adobe-type-tools.github.io/font-tech-notes/pdfs/T1_SPEC.pdf says
in 8.4 First Four Subrs Entries:

"""If Flex or hint replacement is used in a Type 1 font program, the
first four entries in the Subrs array in the Private dictionary must
be assigned charstrings that correspond to the following code
sequences. If neither Flex nor hint replacement is used in the font
program, then this requirement is removed, and the first Subrs
entry may be a normal charstring subroutine sequence. The first
four Subrs entries contain:

Subrs entry number 0:
3 0 callothersubr pop pop setcurrentpoint return

Subrs entry number 1:
0 1 callothersubr return

Subrs entry number 2:
0 2 callothersubr return

Subrs entry number 3:
return

These code sequences must be encoded and encrypted via charstring
encryption in accordance with the rest of the font program. They must
then be included in the program using the
RD–NP format."""

My reading of this is that we don't need explicit code for these because
these strings have to be present in the subrs data.

Also, the explicit code is incorrect in some cases:
* If neither Flex nor hint replacement is used
* For Type 2 CFF data

My motivation is the latter. I could add code to only execute these
conditions if !is_type2, but if we can unconditionally delete this
code, that seems better.