Tests: Add a jpg test input that uses chroma subsampling
I created this jpg by using Adobe Photoshop 2024's Export>Export As…
and picking Format JPG and Quality 1. That was enough to make it use
chroma subsampling.

The image contents were created with a generative fill for the query
"a bowl of pudding, in it a sheet of paper with a mathematical proof
on it. the pudding looks delicious. the sheet of paper is inside the
pudding, not merely on top of it.".

(While the image contents aren't too interesting, the image being
created by generative AI has the side effect of
Content Credentials (Beta) (https://contentcredentials.org/) data
being embedded in the output. The technical spec for that metadata
is as far as I can tell here:
https://c2pa.org/specifications/specifications/1.4/specs/C2PA_Specification.html#_embedding_manifests_into_jpeg
Sadly, it references an ISO standard that can't be easily read
just for the jpeg embedding format already. The spec says it's
an APP11 marker. At least in this file, the app id is 'JP'.
The spec says the marker data is JUMBF containing CBOR. JUMBF is more
ISO nonsense, but CBOR has a nice RFC spec.)

The original generated image was 2100x1500, I scaled it to 25% in the
export dialog.