LibGfx: Add a simple compliance test as suggested in ITU T83
ITU T83 sadly isn't freely available, but the rough summary of what's
suggested in A.1.4 for decoder compliance testing, if I understand it
correctly,  is to have some compressed data, decompress it with the
decoder under test, then compress it back using a reference encoder
that uses doubles everywhere, and then compare the quantized
coefficients that were input data to the quantized coefficients that are
encoder output. The absolute value of this difference should not be
more than 1.

Our JPEGWriter is using doubles for everything, so this uses that
for comparison.

The test sets the quantization matrix to all 1s (i.e. no quantization)
and uses `-1024` in [0][0] and zeros elsewhere. (This is the Y channel
for all black, or all black in a grayscale JPEG.)

Currently, this gets re-encoded as -1016, which is too far away.