AK: Add floor_div() to StdLibExtras.h
It's like ceil_div() except it computes floor instead of ceil.
For positive arguments, `floor_div(a, b)` is the same as
`a / b`.

But if one of the numbers is negative, `a / b` computes the ceil
of the result of the division, while `floor_div(a, b)` keeps computing
the floor.