LibGfx/JPEG2000: Support "selective arithmetic decoder bypass" style
This adds support for the last remaining unsupported code block style.

It's mostly for completionism. I've seen PDFs with the other five
codeblocks in the wild, but I haven't yet seen one using this.

A codeblock (which, simplified, is a small rectangular piece of a single
component of the image) is decoded bitplane by bitplane, most
significant bit to least significant bit.

For each bitplane, three separate passes are run, in this order:

1. A "significance propagation" pass
2. A "magnitude refinement" pass
3. A "cleanup" pass

For the very first bitplane, only the third pass runs. That means
decoding n bitplanes runs 1 + 3*(n-1) passes.

Normally, the data for each of these passes comes from an arithmetic
decoder.

If the "selective arithmetic decoder bypass" style is enabled, the
first 10 passes / 4 bits are decoded normally, but after that,
the bits for the "significance propagation" and "magnitude refinement"
passes are read verbatim from the input bits. The "cleanup" pass still
reads data from an arithmetic decoder.

(This raw input uses the same "stuff bit" procedure as does packet
header decoding: If a byte in the input data is 0xff, the first
bit after it is ignored and must be 0.)

The data that's read verbatim from the input and the data that's
fed to the arithmetic decoder are in independent, byte-aligned
"segments" (which are just chunks for input data).

(A "segment" is different from the "marker segment" concept in the
spec.)

This means:

* The first 10 passes go into one segment, since they all use an
  arithmetic coder
* The next 2 passes go into one segment since these are read verbatim
* The next 1 pass goes into one segment since this uses an arithmetic
  coder again

The last two bullets then repeat.

What makes this a bit complicated that JPEG2000 supports splitting
passes across "layers", roughly packets of data with a fixed number of
decoding passes.

For example, an 8-bit-per-pixel image might use 1 + 3*7 = 22 passes.
(In real life, some of the early bits might be set to zero via a
different mechanism, and some of the late bits might not have to be
coded, so it could also be fewer passes than that.) These 25 passes
could be split across 4 layers which contain 8, 6, 5, and 3 passes.

The first 10 passes are one segment, so the 8 passes from the first
layer and the first 2 passes from the second layer form the 1st segment.
The second layer also contains the 2 passes for the 2nd segment, the 1
pass for the 3rd, and 1 pass of the 4th. The 4th segment is input for
"significance propagation" and "magnitude refinement" and hence needs
data for 2 passes, so the first 1 pass from the third layer also belongs
to the 4th segment. The third layer then also contains 1 pass for the
5th segment, 2 passes for the 6th segment, and 1 pass for the 7th
segment. The fourth layer then contains 2 passes for the 8th segment,
and 1 for the 9th.  The first segment encodes 4 bit planes, and each
pair of segments after tha decodes another 1 bit plane.

When the code block for a layer is read, we know how many passes are in
it. From this, we have to determine how many segments are in this layer
(potentially just partially), and then for each of the segments in this
layer, how many passes are in this segment. With this information, we
can get the number of bytes of raw data in this segment (or,
potentially, segment part). This is done in `read_one_packet_header()`.

The data of one layer is a contiguous blob of data. If a segment is
split over several layers, it needs to be copied into a new contiguous
buffer. So once we've read all the layers, we have to assemble the
split segments into full segments. This is done in
`segments_for_all_layers()`.

This requires a lot of fiddly, lengthy, inelegant code.

The final part is to take the list of segments and alternatingly
read data from arithmetic coders and from bit buffers, depending
on which segment we're in. That's in `decode_code_block()` and
for the most part fairly straightforward. Having all this bit reading
state inline is a bit ugly.

(If the "termination on each pass" code block style is also enabled,
then the data for every single pass is in its own segment, meaning
that the passes using arithmetic decoding all have indepdendent,
byte aligned byte ranges that feed into independent arithmetic
decoders, and "significance propagation" and "magnitude refinement"
passes that read verbatim from the input also both get independent
segments that are aligned to byte boundaries. In this case, things
are much simpler since every pass gets its own segment and they're
never split across layers, and the existing "termination on each pass"
can handle this fine for the most part. `decode_code_block()` has
some slightly tricky logic to read from the right arithmetic
decoder or verbatim buffer, but it's harmless compared with the rest.)

New test images created with:

        ~/Downloads/jasper-build/src/app/jasper --input \
        Tests/LibGfx/*-inputs/*2000/kakadu-lossless-rgba-u8-prog1*.jp2 \
        --output Tests/LibGfx/test-inputs/jpeg2000/jasper-rgba-u8-cbstyle-01-bypass-layers.jp2 \
        -O lazy -O ilyrrates=0.1,0.5,1

        ~/Downloads/jasper-build/src/app/jasper --input \
        Tests/LibGfx/*-inputs/*2000/kakadu-lossless-rgba-u8-prog1*.jp2 \
        --output Tests/LibGfx/test-inputs/jpeg2000/jasper-rgba-u8-cbstyle-05-bypass-termall.jp2 \
        -O lazy -O termall

        ~/Downloads/jasper-build/src/app/jasper --input \
        Tests/LibGfx/*-inputs/*2000/kakadu-lossless-rgba-u8-prog1*.jp2 \
        --output Tests/LibGfx/test-inputs/jpeg2000/jasper-rgba-u8-cbstyle-33-bypass-segsym.jp2 \
        -O lazy -O segsym