Tests: Add another JPEG2000 bypass code block style test image
What's interesting about this one is that it has the the first 10
passes spread out over 3 layers, so that the middle layer contains
a single segment that continues the first segment from the first
layer, and is further continued in the third layer. The second
layer contains not a single complete segment.

The code handles this fine, but this would've caught a bug in
a refactoring that I accidentally introduced locally and that
the tests didn't catch until I added this test case here.

Generated with `jasper -O lazy -O ilyrrates=0.002,0.009,1`.