LibGfx/JBIG2+CCITT: Support decoding MMR-coded JBIG2 graymaps
The main motivation for this is that it's used in page 1 of the test
bitstram in Annex H. Pages 1 and 2 decode to the same bitmap, so
having support for this allows us to compare the two bitmaps, and
we get test coverage for quite a few JBIG2 features that way.

This can also decode amb_2.jb2 from the power jbig2 test dataset.
I'm not aware of any other files using this feature.

An MMR graymap is the only case where the size of the a generic region
is not known in advance, and where the data is immediately followed by
more MMR data.

To make this possible, we add a CCITT function to scan for the
EOFB marker in CCITT group 4 images, and call that from the
graymap decoder.

This has the advantage that it doesn't need any other changes
elsewhere.