Tests/LibGfx: Templatize encode_bitmap()
This allows us to remove the CMYKBitmap overload.

For CMYKBitmap, we'll always go down the `if constexpr`.

Since templates have stricter type checking and all callers pass
NonnullRefPtrs, this changes the function to take a const ref to
NonnullRefPtr. The alternative would be to insert `*`s in front of all
arguments passed to this function.

No behavior change.