Tests/LibGfx: Templatize test_roundtrip()
This allows us to share one version for Bitmap and CMYKBitmap overload.

Since templates have stricter type checking and all callers pass
NonnullRefPtrs, this changes the function to take a const ref to
NonnullRefPtr. The alternative would be to insert `*`s in front of all
arguments passed to this function.

No behavior change.