Tests/LibGfx: Templatize expect_bitmaps_equal()
This allows us to share one version for Bitmap and CMYKBitmap overload.

Since templates have stricter type checking and all callers pass
NonnullRefPtrs, this changes the function to take a const ref to
NonnullRefPtr. The alternative would be to insert `*`s in front of all
arguments passed to this function.

This requires removing `*`s in quite a few places though, and using
`.release_nonnull()` for frames loaded off ImageDecoderPlugins
(which in return requires removing a few `const`s).

No behavior change.